requirements.txt

streamlit
plotly
speedtest-cli
